implemented pair_index_tuples

+++ audalign/__init__.py
@@ -425,7 +425,7 @@ class Audalign:
         against_file_path: str,
         img_width=1.0,
         overlap_ratio=0.5,
-        volume_threshold=215,
+        volume_threshold=215.0,
         plot=False,
     ) -> dict:
         """Recognize target file against against file visually.
@@ -438,7 +438,7 @@ class Audalign:
             against_file_path (str): Recognize against
             img_width (float): width of spectrogram image for recognition
             overlap_ratio (float): overlap of window for matching
-            volume_threshold (int): doesn't find stats for sections with average volume below threshold
+            volume_threshold (float): doesn't find stats for sections with average volume below threshold
             plot (bool): plot the spectrogram of each audio file
 
         Returns
